Refactor Akka Streams based Reactive Client Tests updated
