--- kafka-reactive-system/src/main/scala/org/patricknoir/kafka/reactive/client/actors/KafkaProducerActor.scala
@@ -1,58 +1,0 @@
-package org.patricknoir.kafka.reactive.client.actors
-
-import java.util.concurrent.TimeUnit
-
-import akka.actor.{ Actor, ActorLogging, Props, Status }
-import akka.event.LoggingReceive
-import akka.kafka.ConsumerMessage.CommittableOffset
-import io.circe.Decoder
-import org.apache.kafka.clients.producer.{ KafkaProducer, ProducerRecord }
-import org.patricknoir.kafka.reactive.common.KafkaRequestEnvelope
-import io.circe.generic.auto._
-import io.circe.syntax._
-import org.patricknoir.kafka.reactive.client.actors.KafkaRClientActor.Destination
-import org.patricknoir.kafka.reactive.client.actors.KafkaRRequestActor.KafkaProducerRequest
-import org.patricknoir.kafka.reactive.ex.ProducerException
-
-import scala.collection.JavaConversions._
-import scala.util.Try
-
-/**
- * Created by patrick on 12/07/2016.
- */
-class KafkaProducerActor(producerSettings: Map[String, String]) extends Actor with ActorLogging {
-
-  val producer = new KafkaProducer[String, String](producerSettings)
-
-  def receive = LoggingReceive {
-    case KafkaProducerRequest(envelope: KafkaRequestEnvelope, confirm) =>
-      //      val result = extractDestinationTopic(envelope.destination).map { destTopic =>
-      val record = new ProducerRecord[String, String](envelope.destination.topic, envelope.asJson.noSpaces)
-
-      //TODO: handle the error by throwing a ProducerException to be handled by the suprvisor.
-      //      I need to block: shall I do in a new thread to do not kill the actor performance?
-      //      Risk to lose requests if I wrap in to a future and then I block! Design decision to be made.
-      try {
-        producer.send(record).get(100, TimeUnit.MILLISECONDS)
-        if (confirm) { sender ! (()) }
-      } catch {
-        case err: Throwable =>
-          val producerError = new ProducerException(err)
-          if (confirm) { sender ! Status.Failure(producerError) }
-          throw producerError
-      }
-  }
-
-  private def extractDestinationTopic(destination: String): Option[String] = Try {
-    destination.split(":")(1).split("/")(0)
-  }.toOption
-
-  override def postStop() = {
-    producer.close()
-  }
-
-}
-
-object KafkaProducerActor {
-  def props(producerSettings: Map[String, String]) = Props(new KafkaProducerActor(producerSettings))
-}
