+++ kafka-reactive-system/src/test/scala/org/patricknoir/kafka/reactive/server/ReactiveSystemSpec.scala
@@ -5,11 +5,9 @@ import akka.stream.ActorMaterializer
 import akka.stream.scaladsl._
 import akka.testkit.TestKit
 import org.patricknoir.kafka.reactive.server.dsl._
-import org.patricknoir.kafka.reactive.common.{ KafkaResponseEnvelope, KafkaResponseStatusCode }
-import org.patricknoir.kafka.reactive.common.KafkaRequestEnvelope
+import org.patricknoir.kafka.reactive.common.{ Destination, KafkaRequestEnvelope, KafkaResponseEnvelope, KafkaResponseStatusCode }
 import org.specs2.SpecificationLike
 import io.circe.syntax._
-import org.patricknoir.kafka.reactive.client.actors.KafkaRClientActor.Destination
 
 import scala.concurrent.{ Await, Future }
 import scala.concurrent.duration.Duration
